requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.2.post1
apturl==0.5.2
argcomplete==3.5.2
attrs==24.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
bitarray==2.9.3
bitstring==4.2.3
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
cloudinary==1.42.2
colorama==0.4.4
command-not-found==0.3
contourpy==1.3.2
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
ecdsa==0.19.0
esptool==4.8.1
evdev==1.4.0
exceptiongroup==1.2.2
fastapi==0.115.5
fasteners==0.14.1
fonttools==4.58.5
future==0.18.2
h11==0.14.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
imageio==2.36.1
importlib-metadata==4.6.4
intelhex==2.3.0
jeepney==0.7.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==23.5.0
kiwisolver==1.4.8
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
libvirt-python==8.0.0
llvmlite==0.43.0
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.10.3
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
oauthlib==3.2.0
olefile==0.46
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.3.0
paramiko==2.9.3
pexpect==4.8.0
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
protobuf==3.12.4
protontricks==1.12.1
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.10.2
pydantic_core==2.27.1
PyGObject==3.42.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PyJWT==2.3.0
pymacaroons==0.13.0
PyMatting==1.1.13
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-magic==0.4.24
python-multipart==0.0.19
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reedsolo==1.7.0
referencing==0.35.1
reportlab==3.6.8
requests==2.25.1
rpds-py==0.21.0
scikit-image==0.24.0
scipy==1.14.1
SecretStorage==3.3.1
setproctitle==1.2.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.3.1
starlette==0.41.3
svgwrite==1.4.3
systemd-python==234
tifffile==2024.9.20
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.32.1
vdf==3.4
wadllib==1.3.6
webencodings==0.5.1
xdg==5
xkit==0.0.0
zipp==1.0.0
